mbspipe test: added modes for fast alternating write and read bugfix for correct accounting real transfer buffer size in header; additional checks of expected headers
git-svn-id: https://subversion.gsi.de/dabc/drivers@4612 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ mbspipe/tests/pipe_sync.c
@@ -3,7 +3,7 @@
 
 int f_wait_read(s_pipe_sync* com)
 {
-  int timeout=0;
+  long timeout=0;
   while(com->canread == 0)
   {
     usleep(PIPESYNC_SLEEP);
@@ -15,7 +15,7 @@ int f_wait_read(s_pipe_sync* com)
 
 int f_wait_write(s_pipe_sync* com)
 {
-  int timeout=0;
+  long timeout=0;
   while(com->canwrite == 0)
   {
     usleep(PIPESYNC_SLEEP);
